(PC-29656)[PRO] refactor: Remove GetSiretDataAdapter adapter.
(PC-29656)[PRO] refactor: Remove GetSiretDataAdapter adapter.
