--- pro/src/core/Venue/adapters/getSiretDataAdapter.ts
@@ -1,122 +1,0 @@
-import memoize from 'lodash.memoize'
-
-import { apiAdresse } from 'apiClient/adresse/apiAdresse'
-import { api } from 'apiClient/api'
-import { isErrorAPIError } from 'apiClient/helpers'
-import { unhumanizeSiret } from 'core/Venue/utils'
-import { validateSiret } from 'core/Venue/validate'
-
-type Params = string
-type Payload = {
-  values?: {
-    address: string
-    city: string
-    latitude: number | null
-    longitude: number | null
-    name: string
-    postalCode: string
-    siret: string
-    apeCode: string
-    legalCategoryCode: string
-    banId: string | null
-  }
-}
-type GetSiretDataAdapter = Adapter<Params, Payload, Payload>
-
-const getSiretDataAdapter: GetSiretDataAdapter = async (humanSiret: string) => {
-  const siret = unhumanizeSiret(humanSiret || '')
-  if (humanSiret === '') {
-    return {
-      isOk: true,
-      message: '',
-      payload: {
-        values: {
-          address: '',
-          city: '',
-          latitude: null,
-          longitude: null,
-          name: '',
-          postalCode: '',
-          siret: '',
-          apeCode: '',
-          companyStatus: '',
-          legalUnitStatus: '',
-          legalCategoryCode: '',
-          banId: '',
-        },
-      },
-    }
-  }
-  const error = validateSiret(siret)
-  if (error) {
-    return {
-      isOk: false,
-      message: error,
-      payload: {},
-    }
-  }
-  try {
-    const response = await api.getSiretInfo(siret)
-    /* istanbul ignore next: DEBT, TO FIX */
-    if (!response.active) {
-      return {
-        isOk: false,
-        message: 'SIRET invalide',
-        payload: {},
-      }
-    }
-    const { street, city, postalCode } = response.address
-    const addressData = await apiAdresse.getDataFromAddressParts(
-      street,
-      city,
-      postalCode,
-      1
-    )
-    /* istanbul ignore next: DEBT, TO FIX */
-    if (addressData.length === 0) {
-      return {
-        isOk: false,
-        message: 'Adresse introuvable',
-        payload: {},
-      }
-    }
-    const latitude = addressData[0].latitude
-    const longitude = addressData[0].longitude
-    return {
-      isOk: true,
-      message: `Informations récupéré avec success pour le SIRET: ${humanSiret} :`,
-      payload: {
-        values: {
-          address: response.address.street,
-          city: response.address.city,
-          latitude: latitude,
-          longitude: longitude,
-          name: response.name,
-          postalCode: response.address.postalCode,
-          siret: response.siret,
-          apeCode: response.ape_code,
-          legalCategoryCode: response.legal_category_code,
-          banId: addressData[0].id,
-        },
-      },
-    }
-  } catch (e) {
-    let message = 'Impossible de vérifier le SIRET saisi.'
-    if (isErrorAPIError(e) && e.status === 400) {
-      message = e.body
-      /* istanbul ignore next: DEBT, TO FIX */
-      if (e.body.global) {
-        message = e.body.global[0]
-      }
-    }
-    return {
-      isOk: false,
-      message,
-      payload: {},
-    }
-  }
-}
-
-const getSiretData = memoize(getSiretDataAdapter)
-
-export default getSiretData
